Log messages + OTP.
[IMPROVED]  MP-15: Audit and improve log messages.
[ADDED]     If an element's counter is 0, generate a time-based OTP
            instead.  The OTP changes every 5 minutes.

+++ MasterPassword/MPTypes.h
@@ -82,4 +82,4 @@ NSData *keyHashForKey(NSData *key);
 NSString *NSStringFromMPElementType(MPElementType type);
 NSString *ClassNameFromMPElementType(MPElementType type);
 Class ClassFromMPElementType(MPElementType type);
-NSString *MPCalculateContent(MPElementType type, NSString *name, NSData *key, int16_t counter);
+NSString *MPCalculateContent(MPElementType type, NSString *name, NSData *key, int32_t counter);
